feat: update database models to use BigInteger for ID fields and add add_item method for user inventory management

+++ src/database/models/punishment_log.py
@@ -1,6 +1,6 @@
 from datetime import datetime, timezone
 
-from sqlalchemy import DateTime, Integer, String, Text
+from sqlalchemy import BigInteger, DateTime, String, Text
 from sqlalchemy.orm import Mapped, mapped_column
 
 from database.base import Base
@@ -9,12 +9,12 @@ from database.base import Base
 class PunishmentLog(Base):
     __tablename__: str = "punishment_logs"
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
-    user_id: Mapped[int] = mapped_column(Integer, nullable=False)
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    user_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
     punishment_type: Mapped[str] = mapped_column(String(50), nullable=False)
     reason: Mapped[str] = mapped_column(Text, nullable=False)
-    moderator_id: Mapped[int] = mapped_column(Integer, nullable=False)
-    duration: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    moderator_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
+    duration: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
     created_at: Mapped[datetime] = mapped_column(
         DateTime, nullable=False, default=datetime.now(timezone.utc)
     )
