+++ src/database/models/ticket_channel.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer
+from sqlalchemy import BigInteger
 from sqlalchemy.orm import Mapped, mapped_column
 
 from database.base import Base
@@ -7,6 +7,6 @@ from database.base import Base
 class TicketChannel(Base):
     __tablename__: str = "ticket_channels"
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True)
-    owner_id: Mapped[int] = mapped_column(Integer, nullable=False)
-    category_id: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True)
+    owner_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
+    category_id: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
